relocate router
