+++ package.json
@@ -21,6 +21,7 @@
   "homepage": "https://github.com/lelinho93/backend_byefive#readme",
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
